# internal/recipes/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package ports

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// PostRecipe defines model for PostRecipe.
type PostRecipe struct {
	CookTime     string   `json:"cookTime"`
	Ingredients  []string `json:"ingredients"`
	Instructions []string `json:"instructions"`
	PrepTime     string   `json:"prepTime"`
}

// Recipe defines model for Recipe.
type Recipe struct {
	CookTime     string             `json:"cookTime"`
	Ingredients  []string           `json:"ingredients"`
	Instructions []string           `json:"instructions"`
	PrepTime     string             `json:"prepTime"`
	User         string             `json:"user"`
	UserUuid     openapi_types.UUID `json:"userUuid"`
	Uuid         openapi_types.UUID `json:"uuid"`
}

// Recipes defines model for Recipes.
type Recipes struct {
	Recipes []Recipe `json:"recipes"`
}

// CreateRecipeJSONRequestBody defines body for CreateRecipe for application/json ContentType.
type CreateRecipeJSONRequestBody = PostRecipe

// UpdateRecipeJSONRequestBody defines body for UpdateRecipe for application/json ContentType.
type UpdateRecipeJSONRequestBody = PostRecipe